add 1. fillin 2. get_feedback_user_list 3. feedback 4. statistics

+++ src/main/java/com/example/quiz15/service/impl/QuizServiceImpl.java
@@ -2,7 +2,10 @@ package com.example.quiz15.service.impl;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,19 +13,32 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.example.quiz15.constants.QuestionType;
 import com.example.quiz15.constants.ResCodeMessage;
+import com.example.quiz15.dao.FillinDao;
 import com.example.quiz15.dao.QuestionDao;
 import com.example.quiz15.dao.QuizDao;
 import com.example.quiz15.entity.Question;
 import com.example.quiz15.entity.Quiz;
 import com.example.quiz15.service.ifs.QuizService;
 import com.example.quiz15.vo.BasicRes;
+import com.example.quiz15.vo.FeedbackRes;
+import com.example.quiz15.vo.FeedbackUserRes;
+import com.example.quiz15.vo.FillinReq;
+import com.example.quiz15.vo.OptionCountVo;
+import com.example.quiz15.vo.QuestionAnswerDto;
+import com.example.quiz15.vo.QuestionAnswerVo;
+import com.example.quiz15.vo.QuestionIdAnswerVo;
 import com.example.quiz15.vo.QuestionRes;
 import com.example.quiz15.vo.QuestionVo;
 import com.example.quiz15.vo.QuizCreateReq;
 import com.example.quiz15.vo.QuizUpdateReq;
 import com.example.quiz15.vo.SearchReq;
 import com.example.quiz15.vo.SearchRes;
+import com.example.quiz15.vo.StatisticsRes;
+import com.example.quiz15.vo.StatisticsVo;
+import com.example.quiz15.vo.UserVo;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @Service
@@ -36,6 +52,9 @@ public class QuizServiceImpl implements QuizService {
 
 	@Autowired
 	private QuestionDao questionDao;
+	
+	@Autowired
+	private FillinDao fillinDao;
 
 	/**
 	 * @throws Exception
@@ -136,6 +155,22 @@ public class QuizServiceImpl implements QuizService {
 				ResCodeMessage.SUCCESS.getMessage());
 	}
 
+	private BasicRes checkStatus(LocalDate startDate, boolean isPublished) {
+		// 允許問卷修改的狀態條件: 1. 尚未發佈 2.已發佈+尚未開始
+		// 依照上面條件，程式的寫法如下
+		// if(!isPublished || (isPublished && startDate.isAfter(LocalDate.now())))
+		// 但是因為2個條件式是用 OR (||) 串接，表示只要一個條件成立就會返回成功
+		// 因此只要有比較到 || 後面的條件式，隱含著 isPublished = true，
+		// 上面的 if 條件式可修改成如下
+		if (!isPublished || startDate.isAfter(LocalDate.now())) {
+			// 返回成功表示問卷允許被修改
+			return new BasicRes(ResCodeMessage.SUCCESS.getCode(), //
+					ResCodeMessage.SUCCESS.getMessage());
+		}
+		return new BasicRes(ResCodeMessage.QUIZ_CANNOT_BE_EDITED.getCode(), //
+				ResCodeMessage.QUIZ_CANNOT_BE_EDITED.getMessage());
+	}
+
 	@Transactional(rollbackFor = Exception.class)
 	@Override
 	public BasicRes update(QuizUpdateReq req) throws Exception {
@@ -145,8 +180,10 @@ public class QuizServiceImpl implements QuizService {
 		try {
 			// 1. 檢查 quizId 是否存在
 			int quizId = req.getQuizId();
-			int count = quizDao.getCountByQuizId(quizId);
-			if (count != 1) {
+			// int count = quizDao.getCountByQuizId(quizId);
+			// 不使用 count 數而是取出整筆資料主要是因為後續還會使用到資料庫中的資料
+			Quiz quiz = quizDao.getById(quizId);
+			if (quiz == null) {
 				return new BasicRes(ResCodeMessage.NOT_FOUND.getCode(), //
 						ResCodeMessage.NOT_FOUND.getMessage());
 			}
@@ -155,16 +192,21 @@ public class QuizServiceImpl implements QuizService {
 			if (checkRes.getCode() != 200) { // 不等於 200 表示檢查出有錯誤
 				return checkRes;
 			}
-			// 3. 更新問卷
+			// 3. 檢查原本的問卷狀態(相關欄位的值是存在於DB中)是否可被更新
+			checkRes = checkStatus(quiz.getStartDate(), quiz.isPublished());
+			if (checkRes.getCode() != 200) { // 不等於 200 表示問卷不允許被更新
+				return checkRes;
+			}
+			// 4. 更新問卷
 			int updateRes = quizDao.update(quizId, req.getName(), req.getDescription(), //
 					req.getStartDate(), req.getEndDate(), req.isPublished());
 			if (updateRes != 1) { // 表示資料沒更新成功
 				return new BasicRes(ResCodeMessage.QUIZ_UPDATE_FAILED.getCode(), //
 						ResCodeMessage.QUIZ_UPDATE_FAILED.getMessage());
 			}
-			// 4. 刪除同一張問卷的所有問題
+			// 5. 刪除同一張問卷的所有問題
 			questionDao.deleteByQuizId(quizId);
-			// 5. 檢查問題
+			// 6. 檢查問題
 			List<QuestionVo> questionVoList = req.getQuestionList();
 			for (QuestionVo vo : questionVoList) {
 				// 檢查題目類型與選項
@@ -250,7 +292,7 @@ public class QuizServiceImpl implements QuizService {
 		LocalDate endDate = req.getEndDate() == null ? LocalDate.of(2999, 12, 31) //
 				: req.getEndDate();
 		List<Quiz> list = new ArrayList<>();
-		if(req.isPublished()) {
+		if (req.isPublished()) {
 			list = quizDao.getAllPublished(quizName, startDate, endDate);
 		} else {
 			list = quizDao.getAll(quizName, startDate, endDate);
@@ -266,6 +308,17 @@ public class QuizServiceImpl implements QuizService {
 			return new BasicRes(ResCodeMessage.QUIZ_ID_ERROR.getCode(), //
 					ResCodeMessage.QUIZ_ID_ERROR.getMessage());
 		}
+		Quiz quiz = quizDao.getById(quizId);
+		// 要判斷是否為 null，若不判斷且取得的值是 null 時，後續使用方法會報錯
+		if (quiz == null) {
+			return new BasicRes(ResCodeMessage.NOT_FOUND.getCode(), //
+					ResCodeMessage.NOT_FOUND.getMessage());
+		}
+		// 檢查問卷狀態是否可被更新
+		BasicRes checkRes = checkStatus(quiz.getStartDate(), quiz.isPublished());
+		if (checkRes.getCode() != 200) { // 不等於 200 表示問卷不允許被更新
+			return checkRes;
+		}
 		try {
 			quizDao.deleteById(quizId);
 			questionDao.deleteByQuizId(quizId);
@@ -274,6 +327,280 @@ public class QuizServiceImpl implements QuizService {
 			throw e;
 		}
 		return null;
+	}
+
+	@Transactional(rollbackFor = Exception.class)
+	@Override
+	public BasicRes fillin(FillinReq req) throws Exception {
+		// 檢查填寫的問卷(quiz)
+		// 檢查 1. 是否已發布 2. 當下的日期是否可以填寫(當天是否介於 開始日期和結束日期 之間)
+		int count = quizDao.selectCountById(req.getQuizId(), LocalDate.now());
+		if (count != 1) {
+			return new BasicRes(ResCodeMessage.QUIZ_ID_ERROR.getCode(), //
+					ResCodeMessage.QUIZ_ID_ERROR.getMessage());
+		}
+		// 檢查一個 email 不能再寫同一份問卷
+		count = fillinDao.selectCountByQuizIdAndEmail(req.getQuizId(), req.getEmail());
+		if(count != 0) {
+			return new BasicRes(ResCodeMessage.EMAIL_DUPLICATED.getCode(), //
+					ResCodeMessage.EMAIL_DUPLICATED.getMessage());
+		}
+		// 檢查題目
+		// 檢查 1. 必填但沒有答案 2. 單選但有多個答案 3. 答案跟選項是一樣(答案必須是選項之一)
+		// 取得一張問卷的所有題目
+		List<Question> questionList = questionDao.getQuestionsByQuizId(req.getQuizId());
+		List<QuestionIdAnswerVo> questionAnswerVoList = req.getQuestionAnswerVoList();
+		// 將問題編號和回答轉換成 Map，就是將 QuestionAnswerVo 裡面的2個屬性轉成 Map
+		Map<Integer, List<String>> answerMap = new HashMap<>();
+		for(QuestionIdAnswerVo vo : questionAnswerVoList) {
+			answerMap.put(vo.getQuestionId(), vo.getAnswerList());
+		}
+		// 檢查每一題
+		for(Question question : questionList) {
+			int questionId = question.getQuestionId();
+			String type = question.getType();
+			boolean required = question.isRequired();
+			// 1. 檢查必填但沒有答案 --> 必填但 questionId 沒有在 answerMap 的 key 裡面
+			if(required && !answerMap.containsKey(questionId)) {
+				return new BasicRes(ResCodeMessage.ANSWER_REQUIRED.getCode(), //
+						ResCodeMessage.ANSWER_REQUIRED.getMessage());
+			}
+			// 2. 檢查單選但有多個答案
+			if(type.equalsIgnoreCase(QuestionType.SINGLE.getType())) {
+				List<String> answerList = answerMap.get(questionId);
+				if(answerList.size() > 1) {
+					return new BasicRes(ResCodeMessage.QUESTION_TYPE_IS_SINGLE.getCode(), //
+							ResCodeMessage.QUESTION_TYPE_IS_SINGLE.getMessage());
+				}
+			}
+			// 簡答題沒有選項，跳過該題
+			if(type.equalsIgnoreCase(QuestionType.TEXT.getType())) {
+				continue;
+			}
+			// 3. 比對該題的答案跟選項是否一樣(答案必須是選項之一)
+			String optionsStr = question.getOptions();
+			List<String> answerList = answerMap.get(questionId);
+			for(String answer : answerList) {
+				// 將每個答案比對是否被包含在選項字串中
+				if(!optionsStr.contains(answer)) {
+					return new BasicRes(ResCodeMessage.OPTION_ANSWER_MISMATCH.getCode(), //
+							ResCodeMessage.OPTION_ANSWER_MISMATCH.getMessage());
+				}
+			}
+		}
+		// 存資料: 一題存成一筆資料
+		for(QuestionIdAnswerVo vo : questionAnswerVoList) {
+			// 把 answerList 轉成字串型態
+			try {
+				String str = mapper.writeValueAsString(vo.getAnswerList());
+				fillinDao.insert(req.getQuizId(), vo.getQuestionId(), req.getEmail(), //
+						str, LocalDate.now());
+			} catch (Exception e) {
+				throw e;
+			}
+		}		
+		return new BasicRes(ResCodeMessage.SUCCESS.getCode(), //
+				ResCodeMessage.SUCCESS.getMessage());
+	}
+
+	public BasicRes fillin_test(FillinReq req) {
+		// 檢查填寫的問卷(quiz)
+		// 檢查 1. 是否已發布 2. 當下的日期是否可以填寫(當天是否介於 開始日期和結束日期 之間)
+		int count = quizDao.selectCountById(req.getQuizId(), LocalDate.now());
+		if (count != 1) {
+			return new BasicRes(ResCodeMessage.QUIZ_ID_ERROR.getCode(), //
+					ResCodeMessage.QUIZ_ID_ERROR.getMessage());
+		}
+		// 檢查題目
+		// 檢查 1. 必填但沒有答案 2. 單選但有多個答案 3. 答案跟選項是一樣(答案必須是選項之一)
+		// 取得一張問卷的所有題目
+		List<Question> questionList = questionDao.getQuestionsByQuizId(req.getQuizId());
+		List<QuestionIdAnswerVo> questionAnswerVoList = req.getQuestionAnswerVoList();
+		// questionAnswerVoList 的 size 可能會比 questionList 的 size 少，因為有可能是非必填而沒做答
+		// 因為要知道每一題是否必填、單多選，這樣才能拿填寫的答案來比對
+		// --> 所以 questionList 要當成外層迴圈
+
+		// 先把必填題的 questionId 放到一個 List 中
+		List<Integer> questionIdList = new ArrayList<>();
+		for (Question question : questionList) {
+			if (question.isRequired()) {
+				questionIdList.add(question.getQuestionId());
+			}
+		}
+
+		for (Question question : questionList) {
+			int questionId = question.getQuestionId();
+			String type = question.getType();
+			boolean required = question.isRequired();
+			// 該題是必填 --> 檢查 VoList 中是否有該題的編號存在
+			for (QuestionIdAnswerVo vo : questionAnswerVoList) {
+				int voQuestionId = vo.getQuestionId();
+				// 該題必填但題目編號不包含在 questionIdList，回傳錯誤
+				if (required && !questionIdList.contains(voQuestionId)) {
+					return new BasicRes(ResCodeMessage.ANSWER_REQUIRED.getCode(), //
+							ResCodeMessage.ANSWER_REQUIRED.getMessage());
+				}
+
+				List<String> answerList = vo.getAnswerList();
+				// 檢查相同的 questionId，該題是必填但沒有答案 --> 回傳錯誤
+				if (questionId == voQuestionId && required //
+				// CollectionUtils.isEmpty 有判斷到 null
+				// QuestionAnswerVo 中的 answerList 有給定新的預設值，沒有 mapping 到時會是一個空的 List
+						&& answerList.isEmpty()) {
+					return new BasicRes(ResCodeMessage.ANSWER_REQUIRED.getCode(), //
+							ResCodeMessage.ANSWER_REQUIRED.getMessage());
+				}
+				// 檢查相同的 questionId，單選但有多個答案 --> 回傳錯誤
+				if (questionId == voQuestionId && type == QuestionType.SINGLE.getType() //
+						&& answerList.size() > 1) {
+					return new BasicRes(ResCodeMessage.QUESTION_TYPE_IS_SINGLE.getCode(), //
+							ResCodeMessage.QUESTION_TYPE_IS_SINGLE.getMessage());
+				}
+			}
+		}
+		return null;
+	}	
+
+	@Override
+	public FeedbackUserRes feedbackUserList(int quizId) {
+		if(quizId <= 0) {
+			return new FeedbackUserRes(ResCodeMessage.QUIZ_ID_ERROR.getCode(), //
+					ResCodeMessage.QUIZ_ID_ERROR.getMessage());
+		}
+		List<UserVo> userVoList = fillinDao.selectUserVoList(quizId);
+		return new FeedbackUserRes(ResCodeMessage.SUCCESS.getCode(), //
+				ResCodeMessage.SUCCESS.getMessage(), quizId, userVoList);
+	}
+	
+	@Override
+	public FeedbackRes feedback(int quizId, String email) {
+		if(quizId <= 0) {
+			return new FeedbackRes(ResCodeMessage.QUIZ_ID_ERROR.getCode(), //
+					ResCodeMessage.QUIZ_ID_ERROR.getMessage());
+		}
+		List<QuestionAnswerDto> dtoList = fillinDao.selectQuestionAnswerList(quizId, email);
+		List<QuestionAnswerVo> voList = new ArrayList<>();
+		// 將 QuestionAnswerDto 轉成 QuestionAnswerVo
+		for(QuestionAnswerDto dto : dtoList) {
+			// 將 dot 中的 answerStr 轉換成 List<String>
+			try {
+				List<String> answerList = mapper.readValue(dto.getAnswerStr(), new TypeReference<>() {
+				});
+				// 一對一將 dto 的資料設定到 vo 中 
+				QuestionAnswerVo vo = new QuestionAnswerVo(dto.getQuestionId(), //
+						dto.getQuestion(), dto.getType(), dto.isRequired(), 	answerList);
+				voList.add(vo);
+			} catch (Exception e) {
+				// 不需要 throw 因為沒有使用 @Transactional，就只有取資料而已，所以可以 return 自定義的錯誤資訊
+				return new FeedbackRes(ResCodeMessage.OBJECTMAPPER_PROCESSING_ERROR.getCode(), //
+						ResCodeMessage.OBJECTMAPPER_PROCESSING_ERROR.getMessage());
+			}			
+		}
+		return new FeedbackRes(ResCodeMessage.SUCCESS.getCode(), //
+				ResCodeMessage.SUCCESS.getMessage(), voList);
+	}
+
+	@Override
+	public StatisticsRes statistics(int quizId) {
+		if(quizId <= 0) {
+			return new StatisticsRes(ResCodeMessage.QUIZ_ID_ERROR.getCode(), //
+					ResCodeMessage.QUIZ_ID_ERROR.getMessage());
+		}
+		// 1. 取得問題和作答的資料
+		List<QuestionAnswerDto> dtoList = fillinDao.selectQuestionAnswerList(quizId);
+		// 2. 建立 Map 蒐集 ""選擇題"" 相同題號的所有作答: 作答也是選項，只是 List 中的字串會重複
+		// Map<問題編號, 所有作答>
+		Map<Integer, List<String>> quIdAnswerMap = new HashMap<>();
+		for(QuestionAnswerDto dto : dtoList) {
+			// 跳過簡答題
+			if(dto.getType().equalsIgnoreCase(QuestionType.TEXT.getType())) {
+				continue;
+			}
+			// 將 dot 中的 answerStr 轉換成 List<String>
+			try {
+				List<String> answerList = mapper.readValue(dto.getAnswerStr(), new TypeReference<>() {
+				});
+				if(quIdAnswerMap.containsKey(dto.getQuestionId())) {
+					// Map 的特性是有相同的 key，其對應的 value 會後蓋前；所以 key 若已存在，不能直接 put 值
+					// 若 quIdAnswerMap 中已存在 key，則把 key 對應的 value 取出後，與新的值相加
+					List<String> oldList = quIdAnswerMap.get(dto.getQuestionId());
+					// 把 answerList 加入到原本的 List 中
+					oldList.addAll(answerList);
+					// 再把新的 List 放回到 quIdAnswerMap 中
+					quIdAnswerMap.put(dto.getQuestionId(), oldList);
+				} else {
+					// quIdAnswerMap 中不存在 key，直接把 key-value 放到 quIdAnswerMap 中
+					quIdAnswerMap.put(dto.getQuestionId(), answerList);
+				}				
+			} catch (Exception e) {
+				// 不需要 throw 因為沒有使用 @Transactional，就只有取資料而已，所以可以 return 自定義的錯誤資訊
+				return new StatisticsRes(ResCodeMessage.OBJECTMAPPER_PROCESSING_ERROR.getCode(), //
+						ResCodeMessage.OBJECTMAPPER_PROCESSING_ERROR.getMessage());
+			}
+		}
+		// 3. 取得每題問題的選項
+		List<Question> questionList = questionDao.getQuestionsByQuizId(quizId);
+		// 建立題號和選項次數Vo陣列的 map: Map<題號, 選項次數VoList>
+		Map<Integer, List<OptionCountVo>> quIdOptionVoListMap = new HashMap<>();
+		for(Question question : questionList) {
+			try {
+				// 跳過 簡答題
+				if(question.getType().equalsIgnoreCase(QuestionType.TEXT.getType())) {
+					continue;
+				}
+				// 把選項字串轉換成 List
+				List<String> optionList = mapper.readValue(question.getOptions(), new TypeReference<>() {
+				});
+				List<OptionCountVo> voList = new ArrayList<>();
+				// 把 optionList 中的每個選項轉成 OptionCountVo
+				for(String op : optionList) {
+					// 每個 op 就是一個選項
+					OptionCountVo vo = new OptionCountVo(op, 0);
+					voList.add(vo);
+				}				
+				// 把每題對應的選項 List 放到 map 中
+				// 不需要判斷 quIdOptionListMap 的 key 是否已存在，因為相同問卷，
+				//  其問題編號不會重複，也就是只會有一筆資料而已
+				quIdOptionVoListMap.put(question.getQuestionId(), voList);
+			} catch (Exception e) {
+				return new StatisticsRes(ResCodeMessage.OBJECTMAPPER_PROCESSING_ERROR.getCode(), //
+						ResCodeMessage.OBJECTMAPPER_PROCESSING_ERROR.getMessage());
+			}
+		}
+		// 4. 計算每題每個選項的次數
+		for(Entry<Integer, List<OptionCountVo>> map : quIdOptionVoListMap.entrySet()) {
+			int questionId = map.getKey();
+			List<OptionCountVo> voList = map.getValue();
+			// 1. 從 quIdAnswerMap 中，取得相同 key 對應的 value
+			List<String> answerList = quIdAnswerMap.get(questionId);			
+			for(OptionCountVo vo : voList) {
+				// 紀錄 answerList 的 size，因為每次回圈跑完 answerList 應該都會變更
+				int size = answerList.size();
+				String option = vo.getOption();
+				// 移除所有 answerList 中的符合選項(option) --> 得到新的 answerList
+				answerList.removeAll(List.of(option));
+				// 紀錄新的 answerList 的 size
+				int newSize = answerList.size();
+				// 計算次數並把該值設定回 OptionCountVo 中
+				int count = size - newSize;
+				vo.setCount(count);				
+			}
+		}
+		// 5. 設定 res
+		List<StatisticsVo> statisticsVoList = new ArrayList<>();
+		for(Entry<Integer, List<OptionCountVo>> map : quIdOptionVoListMap.entrySet()) {			
+			// 將從 DB 取得的資料 set 到 vo
+			for(QuestionAnswerDto dto : dtoList) {
+				// 設定相同題號的資料
+				if(map.getKey() == dto.getQuestionId()) {
+					StatisticsVo vo = new StatisticsVo(dto.getQuestionId(), dto.getQuestion(),//
+							dto.getType(), dto.isRequired(), map.getValue());
+					statisticsVoList.add(vo);
+				}				
+			}
+		}
+		return new StatisticsRes(ResCodeMessage.SUCCESS.getCode(), //
+				ResCodeMessage.SUCCESS.getMessage(), statisticsVoList);
 	}
 
 }
